% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_group_data.R
\name{get_feature_group_data}
\alias{get_feature_group_data}
\title{Get feature group data}
\usage{
get_feature_group_data(x)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}
}
\value{
A \code{list} containing the (\verb{$ids}) \code{integer} group identifiers for each
feature, (\verb{$confidence_level}) \code{numeric} confidence_level for each group,
and ($target_trans\verb{) }character` method for transforming targets.
}
\description{
Get the feature group data from a \code{\link[prioritizr:problem]{prioritizr::problem()}} object.
}
\details{
This function is used internally by the robust objective functions
to extract the feature groupings from a \code{\link[prioritizr:problem]{prioritizr::problem()}} based on the
robust constraints.
}
\examples{
# Get planning unit data
pu <- get_sim_pu_raster()

# Get feature data
features <- get_sim_features()

# Define the feature groups,
# Here, we will assign the first 2 features to the group A, and
# the remaining features to the group B
groups <- c(rep("A", 2), rep("B", nlyr(features) - 2))

# Build problem
p <-
  problem(pu, features) |>
  add_robust_min_set_objective() |>
  add_constant_robust_constraints(groups = groups, conf_level = 0.9) |>
  add_relative_targets(0.1) |>
  add_binary_decisions() |>
  add_default_solver(verbose = FALSE)

feature_group_data <- get_feature_group_data(p)
print(feature_group_data)

}
\seealso{
The \code{\link[=add_constant_robust_constraints]{add_constant_robust_constraints()}} function is used to specify feature
groupings.
}
