% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_vic_cost}
\alias{get_vic_cost}
\title{Conservation cost proxy of Victoria, Australia}
\usage{
get_vic_cost()
}
\description{
Load a terra raster of a proxy of conservation cost, from
the Human Footprint Index by Brooke Williams et al. (2020), cropped to
the state of Victoria, Australia.
}
\references{
Williams, B. A., Venter, O., ..., Watson, J. E. M. (2020). Change in Terrestrial Human Footprint Drives Continued Loss of Intact Ecosystems. One Earth, 3, 3, 371-382.
}
