% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_vic_pa}
\alias{get_vic_pa}
\title{Existing protected areas of Victoria, Australia}
\usage{
get_vic_pa()
}
\description{
Load a raster indicating whether the cell is covered by a protected area (1) or not(0).
}
\references{
DCCEEW (2024), Collaborative Australian Protected Areas Database (CAPAD).
}
