% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_vic_species}
\alias{get_vic_species}
\title{Species Occurrence climate projections in Victoria, Australia}
\usage{
get_vic_species()
}
\description{
Load a terra raster object of CMIP6 projections of flora and fauna occurrence in Victoria derived from MaxEnt modelling by Carla Archibald et al. (2024). It was cropped to the state of Victoria, converted to
a presence/ absence metric based on a threshold, and subsetted to only contain
species predictions with adequate model fit (i.e., AUC threshold > 0.7, Boyce threshold > 0.5
and Minimum Area Threshold > 1.0). It was subsetted to contain 17 species of interest.
}
\examples{
vic_species <- get_vic_species()

}
\references{
Archibald, C., Summers, D., Graham, E. and Bryan, B. (2024). Habitat suitability maps
for Australian flora and fauna under CMIP6 climate scenarios. GigaScience, Volume
13, 2024, giae002. \url{https://doi.org/10.1093/gigascience/giae002}
}
