% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions_robust.R
\name{has_robust_constraints}
\alias{has_robust_constraints}
\title{Has robust constraints}
\usage{
has_robust_constraints(x)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}
}
\value{
An invisible \code{logical} value indicating success.
}
\description{
Check if a \code{\link[prioritizr:problem]{prioritizr::problem()}} has robust constraints.
}
