% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_objectives.R
\name{objectives}
\alias{objectives}
\title{Add a robust objective}
\description{
TODO.
}
\details{
The following objectives can be added to a conservation planning
\code{\link[prioritizr:problem]{prioritizr::problem()}}:

\describe{

\item{\code{\link[=add_robust_min_set_objective]{add_robust_min_set_objective()}}}{TODO.}

\item{\code{\link[=add_robust_min_shortfall_objective]{add_robust_min_shortfall_objective()}}}{TODO.}

}
}
\examples{
\dontrun{
# TODO
}
}
\concept{overviews}
