% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.prioritizr-package.R
\docType{package}
\name{robust.prioritizr-package}
\alias{robust.prioritizr-package}
\title{robust.prioritizr: Robust Systematic Conservation Prioritization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Robust systematic conservation prioritization based on chance constrained programming and robust optimization. It enables conservation prioritization users to account of uncertainty in input data from climate change scenario projections, species distribution models or measurement errors. It extends the flexible interface of the 'prioritizr' R package, a systematic conservation prioritization framework, creating robust versions of systematic conservation planning problems that can be solved using exact solvers. Problems can be specified to be fully robust to all data realizations provided, or partially robust, which guarantees that the probability that the constraints are met is higher than the specified confidence level. Fully robust optimization problems are formulated as a "Minimax" linear programming problem, which minimises or maximises the objective under the worst possible data realization. The partially robust optimization problems can be formulated as a Chance Constrained Programming problem (Charnes & Cooper, 1959, \doi{10.1287/mnsc.6.1.73}), which ensures that the constraint is met up to a certain confidence level (probability which the constraint is met), or a Conditional Value-at-Risk optimization problem (Rockafellar & Uryasev, 2000, \doi{10.21314/JOR.2000.038}), where the constraint is met even when taken as the average of the distribution of values that fall above/ below a certain quantile, which also guarantees that the constraint is met up to a certain confidence level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frankiecho/robust.prioritizr}
  \item \url{https://frankiecho.github.io/robust.prioritizr/}
  \item Report bugs at \url{https://github.com/frankiecho/robust.prioritizr/issues}
}

}
\author{
\strong{Maintainer}: Frankie Cho \email{frankie.cho@qut.edu.au} (\href{https://orcid.org/0000-0003-1369-4980}{ORCID})

Authors:
\itemize{
  \item Jeffrey O Hanson \email{jeffrey.hanson@uqconnect.edu.au} (\href{https://orcid.org/0000-0002-4716-6134}{ORCID})
}

}
\keyword{internal}
