% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run example?}
\usage{
run_example()
}
\value{
A \code{logical} value.
}
\description{
Determine if the session is suitable for executing long-running
examples.
}
\details{
This function will return \code{TRUE} if the session is interactive.
Otherwise, it will only return \code{TRUE} if the session does not
have system environmental variables that indicate that the session
is being used for package checks, or for building documentation.
}
\examples{
# should examples be run in current environment?
run_example()

}
