% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_targets.R
\name{transform_targets}
\alias{transform_targets}
\title{Transform targets}
\usage{
transform_targets(target_data, feature_group_data)
}
\arguments{
\item{target_data}{\code{data.frame} with information on the target thresholds.
It should be generated with \code{p$feature_targets()}, where \code{p} is a
\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}

\item{feature_group_data}{\code{data.frame} with information on which
features belong to each feature group. It should be generated with
\code{get_feature_group_data()}.}
}
\value{
A modified version of \code{targets} where the values in the threshold column
are altered according to the target transformation methods specified
by \code{feature_group_data}.
}
\description{
Standardize targets for features that belong to the same feature group.
}
\examples{
# Get planning unit data
pu <- get_sim_pu_raster()

# Get feature data
features <- get_sim_features()

# Define the feature groups,
# Here, we will assign the first 2 features to the group A, and
# the remaining features to the group B
groups <- c(rep("A", 2), rep("B", nlyr(features) - 2))

# Build problem
p <-
  problem(pu, features) |>
  add_robust_min_set_objective() |>
  add_constant_robust_constraints(groups = groups, conf_level = 0.9) |>
  add_relative_targets(0.1) |>
  add_binary_decisions() |>
  add_default_solver(verbose = FALSE)

x <- transform_targets(p$feature_targets(), get_feature_group_data(p))
print(x)

}
