% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vic_cmip6}
\alias{vic_cmip6}
\title{Species Presence/ Absence data in Victoria, Australia}
\format{
\subsection{\code{vic_cmip6}}{

An R object with
\describe{
\item{species}{Wrapped 5km terra SpatRaster object of species presence/ absence under different CIMP6 climate projections}
\item{cost}{Wrapped 5km terra SpatRaster object of conservation cost proxied by the Human Footprint Index}
\item{pa}{Wrapped 5km terra SpatRaster object of protected areas (1 = Protected Area, 0 = Not Protected Area)}
\item{species_details}{Tibble object of the details of the species presence/ absence data, including (a) name: name
of the species layer, (b) species: species name, (c) class: species class, proj: climate projection, auc: Training AUC metric for model fit,
boyce: Spearman correlation (the Boyce index value), threshold: MaxEnt threshold.}
}
}
}
\usage{
vic_cmip6
}
\description{
Contains a table showing information of layers retrieved from \code{get_vic_species()} function of the CMIP6 projections of flora and fauna occurrence in Victoria derived from MaxEnt modelling by Archibald et al. (2024).
}
\details{
Cost data is based on the Human Footprint Index 2013.

Protected area data are from the Collaborative Australian Protected Areas Database (2024),
DCCEEW, Australian Government.

Details of can be found in \href{https://github.com/jeffreyhanson/robust.prioritizr.data}{robust.prioritizr.data}
}
\references{
Archibald, C., Summers, D., Graham, E. and Bryan, B. (2024). Habitat suitability maps
for Australian flora and fauna under CMIP6 climate scenarios. GigaScience, Volume
13, 2024, giae002. \url{https://doi.org/10.1093/gigascience/giae002}
}
\keyword{datasets}
