% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vic_hab_suitability}
\alias{vic_hab_suitability}
\title{Species Presence/ Absence, Cost and Protected Area data in Victoria, Australia}
\format{
\subsection{\code{vic_hab_suitability}}{

An R object with
\describe{
\item{species}{Wrapped 5km terra SpatRaster object of species presence/ absence under different CIMP6 climate projections}
\item{cost}{Wrapped 5km terra SpatRaster object of conservation cost proxied by the Human Footprint Index}
\item{pa}{Wrapped 5km terra SpatRaster object of protected areas (1 = Protected Area, 0 = Not Protected Area)}
\item{species_details}{Tibble object of the details of the species presence/ absence data, including (a) name: name
of the species layer, (b) species: species name, (c) class: species class, proj: climate projection, auc: Training AUC metric for model fit,
boyce: Spearman correlation (the Boyce index value), threshold: MaxEnt threshold.}
...
}
}
}
\usage{
vic_hab_suitability
}
\description{
Contains terra raster objects for solving a systematic conservation planning
problem in Victoria, Australia accounting for uncertainty in species
presence/ absence predictions across future climate change projections.
}
\details{
Projections of species presence/ absence are derived from MaxEnt habitat suitability
maps for Australian flora and fauna under CMIP6 climate scenarios produced by
Archibald et al. (2024). It was cropped to the state of Victoria, converted to
a presence/ absence metric based on a threshold, and subsetted to only contain
species predictions with adequate model fit (i.e., AUC threshold > 0.7, Boyce threshold > 0.5
and Minimum Area Threshold > 1.0).

Cost data is based on the Human Footprint Index 2013.

Protected area data are from the Collaborative Australian Protected Areas Database (2024),
DCCEEW, Australian Government.

Details of can be found in \href{https://github.com/jeffreyhanson/robust.prioritizr.data}{robust.prioritizr.data}
}
\references{
Archibald, C., Summers, D., Graham, E. and Bryan, B. (2024). Habitat suitability maps
for Australian flora and fauna under CMIP6 climate scenarios. GigaScience, Volume
13, 2024, giae002. \url{https://doi.org/10.1093/gigascience/giae002}

DCCEEW (2024), Collaborative Australian Protected Areas Database (CAPAD).
\url{https://www.dcceew.gov.au/environment/land/nrs/science/capad}
}
\keyword{datasets}
