% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_constant_robust_constraints.R
\name{add_constant_robust_constraints}
\alias{add_constant_robust_constraints}
\title{Add constant robust constraints}
\usage{
add_constant_robust_constraints(x, groups, threshold)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}

\item{groups}{\code{character} vector indicating which features
should be grouped together for the purposes of characterizing uncertainty.
In particular, \code{groups} is used to specify a group name for each feature
and features with the same group name will be grouped together.
For example, if some of the features correspond to alternative predictions
for the same species under different scenarios, then these features should
have the same grouping name.}

\item{threshold}{\code{numeric} value describing the level of risk aversion
required for the prioritization (ranging between 0 and 1).
For example, a value of zero corresponds
to a low level of risk aversion, and so the optimization process will not
constrained to be robust against uncertainty. Alternatively, a value of
of one corresponds to a high level of risk version, and so the
optimization process will be highly constrained to be robust against
uncertainty.}
}
\value{
An updated \code{\link[prioritizr:problem]{prioritizr::problem()}} object with the constraint added
to it.
}
\description{
Add robust constraints to a conservation problem to ensure that
the priority areas are robust to uncertainty. In particular, this
function is useful when the level of risk aversion is constant
across all features.
}
\section{Data requirements}{

The robust constraints require that you have multiple alternative
realizations for each biodiversity elements of interest (e.g.,
species, ecosystems, ecosystem services). For example, we might have 5
species of interest. By applying different spatial modeling techniques,
we might have 10 different models for each of the 5 different species.
We can use these models to generate 10 alternative realizations
for each of the 5 species (yielding 50 alternative realizations in total).
To use these data, we would input these 50 alternative realizations
as 50 features when initializing a conservation planning problem
(i.e., \code{\link[prioritizr:problem]{prioritizr::problem()}}) and then use this function to specify which
of the of the features correspond to the same species (based on the feature
groupings parameter).
}

\examples{
\dontrun{
TODO.
}

}
\references{
TODO.
}
\seealso{
See \link{robust_objectives} for an overview of all functions for adding
robust objectives.
}
