% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_robust_min_set_objective.R
\name{add_robust_min_set_objective}
\alias{add_robust_min_set_objective}
\title{Add robust minimum set objective}
\usage{
add_robust_min_set_objective(x)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}
}
\value{
An updated \code{\link[prioritizr:problem]{prioritizr::problem()}} object with the objective added
to it.
}
\description{
TODO.
}
\details{
TODO.
}
\section{Mathematical formulation}{

TODO.
}

\examples{
\dontrun{
TODO.
}

}
\references{
TODO.
}
\seealso{
See \link{robust_objectives} for an overview of all functions for adding
robust objectives.

Other robust_objectives: 
\code{\link{add_robust_min_shortfall_objective}()}
}
\concept{robust_objectives}
