% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_robust_min_shortfall_objective.R
\name{add_robust_min_shortfall_objective}
\alias{add_robust_min_shortfall_objective}
\title{Add robust minimum shortfall objective}
\usage{
add_robust_min_shortfall_objective(x, budget)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}

\item{budget}{\code{numeric} value specifying the maximum expenditure of
the prioritization. For problems with multiple zones, the argument
to \code{budget} can be (i) a single \code{numeric} value to specify a single budget
for the entire solution or (ii) a \code{numeric} vector to specify
a separate budget for each management zone.}
}
\value{
An updated \code{\link[prioritizr:problem]{prioritizr::problem()}} object with the objective added
to it.
}
\description{
TODO.
}
\details{
TODO.
}
\section{Mathematical formulation}{

This objective can be expressed mathematically for a set of planning units
(\eqn{I}{I} indexed by \eqn{i}{i}), a set of features (\eqn{J}{J} indexed
by \eqn{j}{j}), and a set of realizations (\eqn{K}{K} indexed by \eqn{k}{k}) as:

\deqn{\mathit{Minimize} \space  \sum_{j = 1}^{J} w_j \frac{y_j}{t_j} \\
\mathit{subject \space to} \\
\sum_{i = 1}^{I} x_i r_{ijk} + v_{jk} \geq t_j \forall j \in J, k \in K \\
\sum_{i = 1}^{I} x_i c_i \leq B \\
y_j \geq v_{jk} \forall k \in K}{
Minimize sum_j^J wj * (yj / tj) subject to
sum_i^I (xi * rij) + yj >= tj for all j in J &
sum_i^I (xi * ci) <= B
}

Here, \eqn{x_i}{xi} is the \link{decisions} variable (e.g.,
specifying whether planning unit \eqn{i}{i} has been selected (1) or not
(0)), \eqn{r_{ijk}}{rijk} is the amount of feature \eqn{j}{j} in planning
unit \eqn{i}{i} in realization \eqn{k}{k}, \eqn{t_j}{tj} is the representation target for feature
\eqn{j}{j}, \eqn{y_j}{yj} denotes the robust representation shortfall for
the target \eqn{t_j}{tj} for feature \eqn{j}{j} across all realizations \eqn{k}{k},
\eqn{v_{jk}}{vjk} is the shortfall for feature {j} under realization {k}, and \eqn{w_j}{wj} is the
weight for feature \eqn{j}{j} (defaults to 1 for all features; see
\code{\link[=add_feature_weights]{add_feature_weights()}} to specify weights). Additionally,
\eqn{B}{B} is the budget allocated for the solution, \eqn{c_i}{ci} is the
cost of planning unit \eqn{i}{i}. Note that \eqn{y_j}{yj} is a continuous
variable bounded between zero and infinity, and denotes the shortfall
for target \eqn{j}{j}.
}

\examples{
\dontrun{
TODO.
}

}
\references{
TODO.
}
\seealso{
See \link{robust_objectives} for an overview of all functions for adding
robust objectives.

Other robust_objectives: 
\code{\link{add_robust_min_set_objective}()}
}
\concept{robust_objectives}
