% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_robust_min_shortfall_objective.R
\name{add_robust_min_shortfall_objective}
\alias{add_robust_min_shortfall_objective}
\title{Add robust minimum shortfall objective}
\usage{
add_robust_min_shortfall_objective(x, budget)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}

\item{budget}{\code{numeric} value specifying the maximum expenditure of
the prioritization. For problems with multiple zones, the argument
to \code{budget} can be (i) a single \code{numeric} value to specify a single budget
for the entire solution or (ii) a \code{numeric} vector to specify
a separate budget for each management zone.}

\item{method}{the probabilistic constraint formulation method, either \code{Chance} (default) or \code{CondValueAtRisk}. See details.}
}
\value{
An updated \code{\link[prioritizr:problem]{prioritizr::problem()}} object with the objective added
to it.
}
\description{
Solves a minimum shortfall objective that minimises a robust shortfall objective across all features. For each feature, the shortfall objective is quantified based on the distribution of shortfall across all realizations of data. For a fully robust solution, the shortfall metric used for optimization is the maximum shortfall across all realizations of the data. For a partially robust solution (\code{conf_level} < 1), the solution guarantees that the probability that the shortfall is greater than the shortfall metric used for optimisation is less than or equals to 1 - \code{conf_level}.
}
\details{
Minimises an objective based on the total amount of shortfall (difference between the target and the feature representation) weighted across all features, subject to a budget.

For a feature with more than one number of realizations in
the group, it computes the shortfall metric for that feature
based on the following approach:
\itemize{
\item If \code{conf_level = 1}, the shortfall metric is the maximum
of the shortfall across all data realizations in the group
\item If \code{conf_level < 1}, then the shortfall metric is:
\itemize{
\item \code{Method = 'CondValueAtRisk'}: the average of the shortfall
in the set of data realizations that exceed the \code{1-conf_level}
quantile
\item \code{Method = 'Chance'}: the \code{1-conf_level} quantile of the
distribution of shortfall
}
}
}
\section{Mathematical formulation}{

This objective can be expressed mathematically for a set of planning units
(\eqn{I}{I} indexed by \eqn{i}{i}), a set of features (\eqn{J}{J} indexed
by \eqn{j}{j}), and a set of realizations (\eqn{K}{K} indexed
by \eqn{k}{k}) as:

\deqn{\mathit{Minimize} \space  \sum_{j = 1}^{J} w_j y_j \\
\mathit{subject \space to} \\
\Pr_k\{\sum_{i = 1}^{I} x_i r_{ijk} + T_{j} y_j \geq T_j\} > \alpha \qquad \forall j \in J  \\
\sum_{i = 1}^{I} x_i c_i \leq B \\
y_j \geq v_{jk} \qquad \forall k \in K}{'
Minimize sum_j^J wj * yj subject to
sum_i^I (xi * rij) + tj * yj >= tj for all j in J &
sum_i^I (xi * ci) <= B
}

Here, \eqn{x_i}{xi} is the \link{decisions} variable (e.g.,
specifying whether planning unit \eqn{i}{i} has been selected (1) or not
(0)), \eqn{r_{ijk}}{rijk} is the amount of feature \eqn{j}{j} in planning
unit \eqn{i}{i} in realization \eqn{k}{k}, \eqn{t_j}{tj} is the maximum representation target for feature
\eqn{j}{j} across all realizations \eqn{k}{k}, i.e., \eqn{T_j = \max_k(t_{jk})},
\eqn{y_j}{yj} denotes the robust representation shortfall for
the target \eqn{t_j}{tj} for feature \eqn{j}{j} across all realizations \eqn{k}{k},
\eqn{v_{jk}}{vjk} is the shortfall for feature \eqn{j}{j} under realization \eqn{k}{k}, and \eqn{w_j}{wj} is the
weight for feature \eqn{j}{j} (defaults to 1 for all features; see
\code{\link[=add_feature_weights]{add_feature_weights()}} to specify weights). \eqn{\alpha}{\alpha} is the
specified \code{conf_level} confidence level for the uncertain constraint
(as specified in \verb{add_*_robust_constraints}), and ensures that the proportion of
constraints for each feature group \eqn{k}{k} that are held is higher than the
specified confidence level. Additionally,
\eqn{B}{B} is the budget allocated for the solution, \eqn{c_i}{ci} is the
cost of planning unit \eqn{i}{i}. Note that \eqn{y_j}{yj} is a continuous
variable bounded between zero and one, and denotes the shortfall
for target \eqn{j}{j} as a proportion of the total target.

If \code{conf_level = 1} (default), then the probabilistic constraint is simply represented
as:
\deqn{
\sum_{i = 1}^{I} x_i r_{ijk} + T_{j} y_j \geq T_j  \quad \forall \space j \in J, \space k \in K
}

If \code{conf_level < 1}, the probabilistic constraint is parameterised using a Chance Constraint Programming
approach by reformulating the constraint into the following series of constraints:
\deqn{
\sum_{i = 1}^{I} x_i r_{ijk} + T_j y_j + T_j m_{jk} \geq T_{j} \quad \forall \space j \in J, \space k \in K \\
\sum_{k = 1}^{K_j} m_{jk} / K_j \leq 1 - \alpha \quad \forall \space j \in J\\
m \in \{0, 1\}
}

where \eqn{m}{m} is a binary discrete variable indicating whether the constraint has been held or not, and
\eqn{K_j}{K_j} is the number of realizations for the feature \eqn{j}{j}. The auxiliary variable
\eqn{m}{m} is essentially a count of the number of times the probabilistic constraint
has been violated (if \eqn{m_{jk}=1}{m_{jk}=1}), and the problem ensures that the
total amount of times these are violated don't exceed \eqn{1-\alpha}{1-\alpha}.

The "Conditional Value-at-Risk" method found in \code{add_robust_min_set_objective} has not yet
been implemented in this objective function.
}

\examples{
\dontrun{
pu <- get_sim_pu_raster()

# Get feature data
features <- replicate(2, get_sim_features())
features <- rast(features)
names(features) <- paste0("feature_", rep(1:5, 2),
"_scenario_", rep(1:2, each = 5))
relative_budget <- as.numeric(global(pu, 'sum', na.rm = T)) * 0.1

# Get the groups
groups <- rep(paste0("feature_", 1:5), 2)

# Set up prioritizr problem
p <- problem(pu, features) \%>\%
  add_constant_robust_constraints(groups = groups) \%>\%
  add_robust_min_shortfall_objective(budget = relative_budget) \%>\%
  add_relative_targets(0.1) \%>\%
  add_binary_decisions() \%>\%
  add_default_solver()

# Solve the problem
soln <- solve(p)

# Plot the solution
plot(soln)
}

}
\references{
Charnes, A., & Cooper, W. W. (1959). Chance-Constrained Programming. Management Science, 6(1), 73-79.
}
\seealso{
See \link{robust_objectives} for an overview of all functions for adding
robust objectives.

Other robust_objectives: 
\code{\link{add_robust_min_set_objective}()}
}
\concept{robust_objectives}
