% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_robust_min_shortfall_objective.R
\name{add_robust_min_shortfall_objective}
\alias{add_robust_min_shortfall_objective}
\title{Add robust minimum shortfall objective}
\usage{
add_robust_min_shortfall_objective(x, budget)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}

\item{budget}{\code{numeric} value specifying the maximum expenditure of
the prioritization. For problems with multiple zones, the argument
to \code{budget} can be (i) a single \code{numeric} value to specify a single budget
for the entire solution or (ii) a \code{numeric} vector to specify
a separate budget for each management zone.}
}
\value{
An updated \code{\link[prioritizr:problem]{prioritizr::problem()}} object with the objective added
to it.
}
\description{
TODO.
}
\details{
TODO.
}
\section{Mathematical formulation}{

TODO.
}

\examples{
\dontrun{
TODO.
}

}
\references{
TODO.
}
\seealso{
See \link{robust_objectives} for an overview of all functions for adding
robust objectives.

Other robust_objectives: 
\code{\link{add_robust_min_set_objective}()}
}
\concept{robust_objectives}
