% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable_robust_constraints.R
\name{add_variable_robust_constraints}
\alias{add_variable_robust_constraints}
\title{Add variable robust constraints}
\usage{
add_variable_robust_constraints(x, data)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}

\item{data}{\code{\link[tibble:tibble]{tibble::tibble()}} data frame containing information the feature
groupings and their desired level of risk aversion. See the Data format
section for details. Also, see the Examples section for example usage.}
}
\value{
An updated \code{\link[prioritizr:problem]{prioritizr::problem()}} object with the constraint added
to it.
}
\description{
Add robust constraints to a conservation problem to ensure that
the priority areas are robust to uncertainty. In particular, this
function is useful when the level of risk aversion is different for
different features. For example, this function may be especially
useful when it is important to ensure that a prioritization is
highly robust to uncertainty in the spatial distribution of threatened
species, and only moderately robust to uncertainty in the spatial
distribution of widespread species.
}
\section{Data format}{

The \code{data} argument must be a \code{\link[tibble:tibble]{tibble::tibble()}} data frame that has
information on the feature groupings and their levels of risk aversion.
Here, each row corresponds to a different feature group and
columns contain information about the groups.
In particular, it has the following columns.
\describe{
\item{features}{
A \code{list} column with the names of the features that belong to each group.
In particular, if a particular set of features should belong to the same
group, then they should be stored in the same element of this column.
}
\item{threshold}{
A \code{numeric} column with values that describe the level of risk aversion
associated with each feature group (ranging between 0 and 1).
For example, a value of zero corresponds
to a low level of risk aversion, and so the optimization process will not
constrained to be robust against uncertainty for that particular group.
Alternatively, a value of of one corresponds to a high level of risk
version, and so the optimization process will be highly constrained to be
robust against uncertainty for that particular group.
}
}
}

\section{Data requirements}{

The robust constraints require that you have multiple alternative
realizations for each biodiversity elements of interest (e.g.,
species, ecosystems, ecosystem services). For example, we might have 5
species of interest. By applying different spatial modeling techniques,
we might have 10 different models for each of the 5 different species.
We can use these models to generate 10 alternative realizations
for each of the 5 species (yielding 50 alternative realizations in total).
To use these data, we would input these 50 alternative realizations
as 50 features when initializing a conservation planning problem
(i.e., \code{\link[prioritizr:problem]{prioritizr::problem()}}) and then use this function to specify which
of the of the features correspond to the same species (based on the feature
groupings parameter).
}

\examples{
\dontrun{
TODO.
}

}
\references{
TODO.
}
\seealso{
See \link{robust_objectives} for an overview of all functions for adding
robust objectives.
}
