% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_vic_study_area}
\alias{get_vic_study_area}
\title{Study area polygon of Victoria, Australia}
\usage{
get_vic_study_area()
}
\description{
Load an R sf object of the state of Victoria, Australia, defining the study area
of the Victoria case study.
}
