% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_robust_constraints.R
\name{has_robust_constraints}
\alias{has_robust_constraints}
\title{Has robust constraints?}
\usage{
has_robust_constraints(x)
}
\arguments{
\item{x}{\code{\link[prioritizr:problem]{prioritizr::problem()}} object.}
}
\value{
An invisible \code{logical} value indicating success.
}
\description{
Check if a \code{\link[prioritizr:problem]{prioritizr::problem()}} has robust constraints.
}
\examples{
# Get planning unit data
pu <- get_sim_pu_raster()

# Get feature data
features <- get_sim_features()

# Define the feature groups,
# Here, we will assign the first 2 features to the group A, and
# the remaining features to the group B
groups <- c(rep("A", 2), rep("B", nlyr(features) - 2))

# Build problem
p <-
  problem(pu, features) |>
  add_robust_min_set_objective() |>
  add_constant_robust_constraints(groups = groups, conf_level = 0.9) |>
  add_relative_targets(0.1) |>
  add_binary_decisions() |>
  add_default_solver(verbose = FALSE)

robust_constraints <- has_robust_constraints(p)
print(robust_constraints)

}
