% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{robust.prioritizr}
\alias{robust.prioritzr}
\alias{robust.prioritzr-package}
\alias{robust.prioritizr}
\alias{robust.prioritizr-package}
\title{robust.prioritizr: Spatial conservation planning under uncertainty}
\description{
TODO.
}
\details{
TODO.
}
\section{Citation}{

TODO.
}

\references{
TODO.
}
\seealso{
TODO.
}
\author{
TODO.
}
