% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{robust.prioritizr}
\alias{robust.prioritizr}
\alias{robust.prioritizr-package}
\title{robust.prioritizr: Robust Systematic Conservation Prioritization in R}
\description{
The \pkg{robust.prioritizr} R package provides tools for building and solving
robust systematic conservation prioritization problems. It extends the
\pkg{prioritizr} package to account for uncertainty in the input data.
This is particularly useful when working with data that is subject to
change, such as species distribution models under climate change scenarios.
}
\details{
The main functions provided by this package are:
\itemize{
\item \code{\link[=add_robust_min_set_objective]{add_robust_min_set_objective()}}
\item \code{\link[=add_robust_min_shortfall_objective]{add_robust_min_shortfall_objective()}}
\item \code{\link[=add_constant_robust_constraints]{add_constant_robust_constraints()}}
\item \code{\link[=add_variable_robust_constraints]{add_variable_robust_constraints()}}
}

For more information on using this package, see the vignettes:
\code{vignette("robust.prioritizr", package = "robust.prioritizr")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frankiecho/robust.prioritizr}
  \item \url{https://frankiecho.github.io/robust.prioritizr/}
  \item Report bugs at \url{https://github.com/frankiecho/robust.prioritizr/issues}
}

}
\author{
\strong{Maintainer}: Frankie Cho \email{frankie.cho@qut.edu.au} (\href{https://orcid.org/0000-0003-1369-4980}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jeffrey O Hanson \email{jeffrey.hanson@uqconnect.edu.au} (\href{https://orcid.org/0000-0002-4716-6134}{ORCID})
}

}
