% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vic_cmip6}
\alias{vic_cmip6}
\title{Species Presence/ Absence data in Victoria, Australia}
\format{
\subsection{\code{vic_cmip6}}{

An tibble with
\describe{
\item{id}{ID of the layer}
\item{name}{Full name of the layer}
\item{species}{Name of the species}
\item{class}{Class of the species}
\item{proj}{Climate projection and timestep}
\item{auc}{AUC metric from the model fitting process}
\item{boyce}{Boyce metric from the model fitting process}
\item{threshold}{Threshold}
\item{scenario}{Climate scenario, can be: "historic_baseline" (historic baseline), ssp126 (SSP1-RCP2.6), ssp245 (SSP2-RCP4.5), ssp370 (SSP3-RCP7.0), or ssp585 (SSP5-RCP8.5)}
\item{timestep}{Timestep (year)}
\item{sum}{Sum of the total number of cells in Victoria with historic/ projected presence under the climate scenario in that timestep}
}
}
}
\usage{
vic_cmip6
}
\description{
Contains a table showing information of layers retrieved from \code{get_vic_species()} function of the CMIP6 projections of flora and fauna occurrence in Victoria derived from MaxEnt modelling by Archibald et al. (2024).
}
\details{
See the \code{vic-cons-planning.Rmd} vignette for a demonstration of its usage.

Details of can be found in \href{https://github.com/jeffreyhanson/robust.prioritizr.data}{robust.prioritizr.data}
}
\references{
Archibald, C., Summers, D., Graham, E. and Bryan, B. (2024). Habitat suitability maps
for Australian flora and fauna under CMIP6 climate scenarios. GigaScience, Volume
13, 2024, giae002. \url{https://doi.org/10.1093/gigascience/giae002}
}
\keyword{datasets}
